Add default implementations to deprecated methods of BuilableItem and Item.
Currently the interface requires the API user to implement already deprecated methods.
It does not make much sense, and the API could be simplified.
