+++ core/src/main/java/hudson/model/Item.java
@@ -25,6 +25,8 @@
 package hudson.model;
 
 import hudson.Functions;
+import hudson.Util;
+import jenkins.model.Jenkins;
 import jenkins.util.SystemProperties;
 import hudson.security.PermissionScope;
 import jenkins.util.io.OnMaster;
@@ -131,18 +133,28 @@ public interface Item extends PersistenceRoot, SearchableModelObject, AccessCont
     /**
      * Gets the relative name to this item from the specified group.
      *
+     * @param g
+     *      The ItemGroup instance used as context to evaluate the relative name of this AbstractItem
+     * @return
+     *      The name of the current item, relative to p. Nested ItemGroups are separated by {@code /} character.
      * @since 1.419
      * @return
-     *      String like "../foo/bar"
+     *      String like "../foo/bar".
+     *      {@code null} if item parents is not an {@link ItemGroup}.
      */
-    String getRelativeNameFrom(ItemGroup g);
+    default String getRelativeNameFrom(ItemGroup g) {
+        return Functions.getRelativeNameFrom(this, g);
+    }
 
     /**
      * Short for {@code getRelativeNameFrom(item.getParent())}
      *
      * @since 1.419
      */
-    String getRelativeNameFrom(Item item);
+    default String getRelativeNameFrom(Item item)  {
+        return getRelativeNameFrom(item.getParent());
+
+    }
 
     /**
      * Returns the URL of this item relative to the context root of the application.
@@ -180,7 +192,12 @@ public interface Item extends PersistenceRoot, SearchableModelObject, AccessCont
      *      (even this won't work for the same reason, which should be fixed.)
      */
     @Deprecated
-    String getAbsoluteUrl();
+    default String getAbsoluteUrl() {
+        String r = Jenkins.getInstance().getRootUrl();
+        if(r==null)
+            throw new IllegalStateException("Root URL isn't configured yet. Cannot compute absolute URL.");
+        return Util.encode(r+getUrl());
+    }
 
     /**
      * Called right after when a {@link Item} is loaded from disk.
@@ -207,7 +224,9 @@ public interface Item extends PersistenceRoot, SearchableModelObject, AccessCont
      *
      * @since 1.374
       */
-    default void onCreatedFromScratch() {}
+    default void onCreatedFromScratch() {
+        // do nothing by default
+    }
 
     /**
      * Save the settings to a file.
